--- src/shaders.rs
@@ -1,6 +1,0 @@
-// pub static FRAG_SHADER_SRC: &'static str = "
-// #version 330 core
-// const int f=80;const float v=1920,y=1080;uniform vec4 sp[(f+2)*2];uniform sampler2D terrain;in vec4 gl_FragCoord;out vec4 fragColor;const float e=99999.;vec2 p[4]=vec2[4](normalize(vec2(.23,.65)),normalize(vec2(.83,-.26)),normalize(vec2(.13,-.83)),normalize(vec2(-.2,.55)));vec3 t(vec3 v){float f=0.,i=0.;for(int y=0;y<4;y++){float e=v.x*p[y].x+v.z*p[y].y;e=e*(4.-float(y)*.51013)+sp[162].z;float s=1./(float(y)+1.);f+=s*sin(e-.3*cos(e));i+=s*cos(e-.3*sin(e));}return vec3(f,0.,i);}bool t(vec3 f,vec3 e,vec3 v,float y,out float i){vec3 z=v-e;float s=dot(z,f);if(s<0.)return false;else{float n=length(z),o=n*n-s*s;if(o>y)return false;else{float t=sqrt(y-o);i=s-t;return true;}}}float t(vec2 f,out float v){vec4 s=texture(terrain,f/512.);v=s.y;return s.x*60.-12.1;}void t(vec3 f,vec3 y,out float v,out float i){vec3 s=(vec3(0.)-f)/y,o=(vec3(512.)-f)/y;v=max(min(s.x,o.x),min(s.z,o.z));i=min(max(s.x,o.x),max(s.z,o.z));}bool t(vec3 v,vec3 f,out float i,out vec3 y,out vec3 o,out float e,out float s){float z,n;t(v,f,z,n);if(n<z)return false;float x=max(0.,z);v=v+x*f;vec2 m,r,d;d=sign(f.xz);r=1./f.xz*d;m=r*(max(d,0)-fract(v.xz)*d);if(any(isinf(m)||isinf(r)))return false;vec2 c=floor(v.xz);float p=v.y,u=t(c,s);e=1.2;for(i=0.;i<n-x;){vec2 k=vec2(float(m.x<m.y),float(m.x>=m.y));i=dot(m,k);float g=v.y+f.y*i;c=c+d*k;m=m+r*k;if(u>g)return y=vec3(.2,.071,.01)+step(27,u)*vec3(.1,-.06,0),o=vec3(0,1.,0),i=(u-v.y)/f.y,true;u=t(c,s);if(u>g){e=1.5;y=vec3(.2,.2,.2)+step(40,u)*vec3(0.,-.03,-.1);if(s==1.){float a=25.31;if(g<a)s=0.;}o=vec3(-d.x*k.x,0,-d.y*k.y);return true;}}return false;}float t(float f,vec3 v,vec3 o){float i=(1.-f)/(1.+f);i*=i;float y=-dot(v,o),s=1.-y,e=i+(1.-i)*s*s*s*s*s;return e;}const vec3 i=vec3(5e-06,1.5e-05,.00027)*15.,o=vec3(.00015,.00015,.00027)*15.;vec3 n(float f){return exp(-f*(i+o));}vec3 n(float f,float y){float s=.0003/16.*3.14159*(1.+y*y);vec3 v=1./(i+o)*(1.-exp(-f*o));float e=.476;vec3 x=vec3(.002,.0008,.0002)*(1.-e)*(1.-e)/(12.5664*pow(1.+e*e-2.*e*y,1.5));float k=20./(i.x+o.x)*(1.-exp(-f*o.x));return s*v+x*k;}void main(){vec3 s=normalize(vec3(1.,1.1,1.));vec2 i=2.*(gl_FragCoord.xy/y)-vec2(v/y,1.);vec4 o=cos(sp[161]),d=sin(sp[161]);mat3 x=mat3(o.y,0,-d.y,-d.x*d.y,o.x,-d.x*o.y,o.x*d.y,d.x,o.y*o.x);vec3 m=x*vec3(i,-2.),z=sp[160].xyz;m+=z;vec3 r=normalize(m-z);float k=1.;vec3 c=vec3(0);for(int u=2;u>0;u--){vec3 p,g,a,b;float w,l=0.,C=e;for(int F=0;F<f;F++){float D;if(t(r,z,sp[F*2].xyz,sp[F*2].w,D)){if(D<C)C=D,a=z+C*r,g=normalize(a-sp[F*2].xyz),b=sp[F*2+1].xyz,w=sp[F*2+1].w,l=t(w,g,r);}}float F;vec3 D,q;float h;if(t(z,r,F,D,q,w,h)){if(F<C)C=F,b=D,g=q,a=z+r*C*.9999,l=t(w,g,r);}else if(r.y<0.&&C==e)C=(-10.5-z.y)/r.y;F=(-.5-z.y)/r.y;if(r.y<0.&&F<=C){a=z+r*F*.9999;g=vec3(0.,1.f,0.f)+t(a)*.03/F;g=normalize(g);l=t(1.1,g,r);vec3 Z=refract(r,g,1.-l);b=vec3(.05,.05,.15);if(t(a,Z*100,F,D,q,w,h))b+=D*exp(-F*40.);}p=reflect(normalize(r),g);if(C>=e||u==1){c+=n(C,dot(s,r))*k;break;}bool Z=t(a,s,F,D,q,w,h);if(!Z){for(int Y=0;Y<f;Y++){if(t(s,a,sp[Y*2].xyz,sp[Y*2].w,F)){Z=true;break;}}}if(!Z){float Y=dot(s,g);vec3 X=normalize(s-r);float W=pow(dot(g,X),121.);W=clamp(W,0.,1.);D=vec3(W)+b*Y;}else D=b*.02;D*=n(C);D+=n(C,dot(s,r));c+=D*k*(1.-l);k=k*l;r=p;z=a;}fragColor=vec4(pow(c,vec3(1./2.2)),1.);}\0\0";
-//const int f=80;const float v=1280,y=720;uniform vec4 sp[(f+2)*2];uniform sampler2D terrain;in vec4 gl_FragCoord;out vec4 fragColor;const float e=99999.;vec2 p[4]=vec2[4](normalize(vec2(.23,.65)),normalize(vec2(.83,-.26)),normalize(vec2(.13,-.83)),normalize(vec2(-.2,.55)));vec3 t(vec3 v){float f=0.,i=0.;for(int y=0;y<4;y++){float e=v.x*p[y].x+v.z*p[y].y;e=e*(4.-float(y)*.51013)+sp[162].z;float s=1./(float(y)+1.);f+=s*sin(e-.3*cos(e));i+=s*cos(e-.3*sin(e));}return vec3(f,0.,i);}bool t(vec3 f,vec3 e,vec3 v,float y,out float i){vec3 z=v-e;float s=dot(z,f);if(s<0.)return false;else{float n=length(z),o=n*n-s*s;if(o>y)return false;else{float t=sqrt(y-o);i=s-t;return true;}}}float t(vec2 f,out float v){vec4 s=texture(terrain,f/512.);v=s.y;return s.x*60.-12.1;}void t(vec3 f,vec3 y,out float v,out float i){vec3 s=(vec3(0.)-f)/y,o=(vec3(512.)-f)/y;v=max(min(s.x,o.x),min(s.z,o.z));i=min(max(s.x,o.x),max(s.z,o.z));}bool t(vec3 v,vec3 f,out float i,out vec3 y,out vec3 o,out float e,out float s){float z,n;t(v,f,z,n);if(n<z)return false;float x=max(0.,z);v=v+x*f;vec2 m,r,d;d=sign(f.xz);r=1./f.xz*d;m=r*(max(d,0)-fract(v.xz)*d);if(any(isinf(m)||isinf(r)))return false;vec2 c=floor(v.xz);float p=v.y,u=t(c,s);e=1.2;for(i=0.;i<n-x;){vec2 k=vec2(float(m.x<m.y),float(m.x>=m.y));i=dot(m,k);float g=v.y+f.y*i;c=c+d*k;m=m+r*k;if(u>g)return y=vec3(.2,.071,.01)+step(27,u)*vec3(.1,-.06,0),o=vec3(0,1.,0),i=(u-v.y)/f.y,true;u=t(c,s);if(u>g){e=1.5;y=vec3(.2,.2,.2)+step(40,u)*vec3(0.,-.03,-.1);if(s==1.){float a=25.31;if(g<a)s=0.;}o=vec3(-d.x*k.x,0,-d.y*k.y);return true;}}return false;}float t(float f,vec3 v,vec3 o){float i=(1.-f)/(1.+f);i*=i;float y=-dot(v,o),s=1.-y,e=i+(1.-i)*s*s*s*s*s;return e;}const vec3 i=vec3(5e-06,1.5e-05,.00027)*15.,o=vec3(.00015,.00015,.00027)*15.;vec3 n(float f){return exp(-f*(i+o));}vec3 n(float f,float y){float s=.0003/16.*3.14159*(1.+y*y);vec3 v=1./(i+o)*(1.-exp(-f*o));float e=.476;vec3 x=vec3(.002,.0008,.0002)*(1.-e)*(1.-e)/(12.5664*pow(1.+e*e-2.*e*y,1.5));float k=20./(i.x+o.x)*(1.-exp(-f*o.x));return s*v+x*k;}void main(){vec3 s=normalize(vec3(1.,1.1,1.));vec2 i=2.*(gl_FragCoord.xy/y)-vec2(v/y,1.);vec4 o=cos(sp[161]),d=sin(sp[161]);mat3 x=mat3(o.y,0,-d.y,-d.x*d.y,o.x,-d.x*o.y,o.x*d.y,d.x,o.y*o.x);vec3 m=x*vec3(i,-2.),z=sp[160].xyz;m+=z;vec3 r=normalize(m-z);float k=1.;vec3 c=vec3(0);for(int u=2;u>0;u--){vec3 p,g,a,b;float w,l=0.,C=e;for(int F=0;F<f;F++){float D;if(t(r,z,sp[F*2].xyz,sp[F*2].w,D)){if(D<C)C=D,a=z+C*r,g=normalize(a-sp[F*2].xyz),b=sp[F*2+1].xyz,w=sp[F*2+1].w,l=t(w,g,r);}}float F;vec3 D,q;float h;if(t(z,r,F,D,q,w,h)){if(F<C){if(h==1.){float Z=25.91;vec3 Y=(z-vec3(279.,Z,285.))*512.;float X;if(t(Y,r*512,X,D,q,w,h))C=F+X/512.,b=D,b.z=b.z*2.;}else C=F,b=D;g=q;a=z+r*C*.9999;l=t(w,g,r);}}else if(r.y<0.&&C==e)C=(-10.5-z.y)/r.y;F=(-.5-z.y)/r.y;if(r.y<0.&&F<=C){a=z+r*F*.9999;g=vec3(0.,1.f,0.f)+t(a)*.03/F;g=normalize(g);l=t(1.1,g,r);vec3 Z=refract(r,g,1.-l);b=vec3(.05,.05,.15);if(t(a,Z*100,F,D,q,w,h))b+=D*exp(-F*40.);}p=reflect(normalize(r),g);if(C>=e||u==1){c+=n(C,dot(s,r))*k;break;}bool Z=t(a,s,F,D,q,w,h);if(!Z){for(int Y=0;Y<f;Y++){if(t(s,a,sp[Y*2].xyz,sp[Y*2].w,F)){Z=true;break;}}}if(!Z){float Y=dot(s,g);vec3 X=normalize(s-r);float W=pow(dot(g,X),121.);W=clamp(W,0.,1.);D=vec3(W)+b*Y;}else D=b*.02;D*=n(C);D+=n(C,dot(s,r));c+=D*k*(1.-l);k=k*l;r=p;z=a;}fragColor=vec4(pow(c,vec3(1./2.2)),1.);}\0\0";
-//const int f=80;const float v=1280,z=720;uniform vec4 sp[(f+2)*2];uniform sampler2D terrain;in vec4 gl_FragCoord;out vec4 fragColor;const float e=99999.;vec2 p[4]=vec2[4](normalize(vec2(.23,.65)),normalize(vec2(.83,-.26)),normalize(vec2(.13,-.83)),normalize(vec2(-.2,.55)));vec3 t(vec3 v){float f=0.,i=0.;for(int z=0;z<4;z++){float e=v.x*p[z].x+v.z*p[z].y;e=e*(4.-float(z)*.51013)+sp[162].z;float s=1./(float(z)+1.);f+=s*sin(e-.3*cos(e));i+=s*cos(e-.3*sin(e));}return vec3(f,0.,i);}bool t(vec3 f,vec3 e,vec3 v,float z,out float i){vec3 y=v-e;float s=dot(y,f);if(s<0.)return false;else{float n=length(y),o=n*n-s*s;if(o>z)return false;else{float t=sqrt(z-o);i=s-t;return true;}}}float t(vec2 f,out float v){vec4 s=texture(terrain,f/512.);v=s.y;return s.x*60.-12.1;}void t(vec3 f,vec3 e,out float v,out float i){vec3 s=(vec3(0.)-f)/e,o=(vec3(512.)-f)/e;v=max(min(s.x,o.x),min(s.z,o.z));i=min(max(s.x,o.x),max(s.z,o.z));}bool t(vec3 v,vec3 f,out float i,out vec3 e,out vec3 o,out float z,out float s){float y,n;t(v,f,y,n);if(n<y)return false;float x=max(0.,y);v=v+x*f;vec2 m,r,d;d=sign(f.xz);r=1./f.xz*d;m=r*(max(d,0)-fract(v.xz)*d);if(any(isinf(m)||isinf(r)))return false;vec2 c=floor(v.xz);float p=v.y,u=t(c,s);z=0.;for(i=0.;i<n-x;){vec2 k=vec2(float(m.x<m.y),float(m.x>=m.y));i=dot(m,k);float g=v.y+f.y*i;c=c+d*k;m=m+r*k;if(u>g)return e=vec3(.2,.071,.01)+step(27,u)*vec3(.1,-.06,0),z=1.2,o=vec3(0,1.,0),i=(u-v.y)/f.y,true;u=t(c,s);if(u>g){z=1.5;e=vec3(.2,.2,.2)+step(40,u)*vec3(0.,-.03,-.1);if(s==1.){float a=25.31;if(g<a)s=0.;}o=vec3(-d.x*k.x,0,-d.y*k.y);return true;}}return false;}float t(float f,vec3 v,vec3 o){float e=(1.-f)/(1.+f);e*=e;float z=-dot(v,o),s=1.-z,i=e+(1.-e)*s*s*s*s*s;return i;}const vec3 i=vec3(5e-06,1.5e-05,.00027)*15.,o=vec3(.00015,.00015,.00027)*15.;vec3 n(float f){return exp(-f*(i+o));}vec3 n(float f,float s){float v=.0003/16.*3.14159*(1.+s*s);vec3 e=1./(i+o)*(1.-exp(-f*o));float z=.476;vec3 x=vec3(.002,.0008,.0002)*(1.-z)*(1.-z)/(12.5664*pow(1.+z*z-2.*z*s,1.5));float k=20./(i.x+o.x)*(1.-exp(-f*o.x));return v*e+x*k;}void main(){vec3 s=normalize(vec3(1.,1.1,1.));vec2 i=2.*(gl_FragCoord.xy/z)-vec2(v/z,1.);vec4 o=cos(sp[161]),d=sin(sp[161]);mat3 x=mat3(o.y,0,-d.y,-d.x*d.y,o.x,-d.x*o.y,o.x*d.y,d.x,o.y*o.x);vec3 m=x*vec3(i,-2.),y=sp[160].xyz;m+=y;vec3 r=normalize(m-y);float k=1.;vec3 c=vec3(0);for(int u=2;u>0;u--){vec3 p,g,a,b;float w,l=0.,C=e;for(int F=0;F<f;F++){float D;if(t(r,y,sp[F*2].xyz,sp[F*2].w,D)){if(D<C)C=D,a=y+C*r,g=normalize(a-sp[F*2].xyz),b=sp[F*2+1].xyz,w=sp[F*2+1].w,l=t(w,g,r);}}float F;vec3 D,q;float h;if(t(y,r,F,D,q,w,h)){if(F<C){if(h==1.){float Z=25.91;vec3 Y=(y-vec3(386.,Z,447.))*512.;float X;if(t(Y,r*512,X,D,q,w,h))C=F+X/512.,b=D,b.z=b.z*2.;else{C=F*2.02;y=y+r*C;continue;}}else C=F,b=D;g=q;a=y+r*C*.9999;l=t(w,g,r);}}F=(-.5-y.y)/r.y;if(r.y<0.&&F<=C){a=y+r*F*.9999;g=vec3(0.,1.f,0.f)+t(a)*.01;g=normalize(g);l=t(1.1,g,r);vec3 Z=refract(r,g,1.-l);b=vec3(.05,.05,.15);if(t(a,Z*100,F,D,q,w,h))b+=D*exp(-F*40.);}p=reflect(normalize(r),g);if(C>=e){c+=n(C,dot(s,r))*k;break;}bool Z=t(a,s,F,D,q,w,h);if(!Z){for(int Y=0;Y<f;Y++){if(t(s,a,sp[Y*2].xyz,sp[Y*2].w,F)){Z=true;break;}}}if(!Z){float Y=dot(s,g);vec3 X=normalize(s-r);float W=pow(dot(g,X),121.);W=clamp(W,0.,1.);D=vec3(W)+b*Y;}else D=b*.02;D*=n(C);D+=n(C,dot(s,r));c+=D*k*(1.-l);k=k*l;r=p;y=a;}fragColor=vec4(pow(c,vec3(1./2.2)),1.);}\0\0";
-//const int f=80;const float v=1280,y=720;uniform vec4 sp[(f+2)*2];uniform sampler2D terrain;in vec4 gl_FragCoord;out vec4 fragColor;const float e=99999.;vec2 p[4]=vec2[4](normalize(vec2(.23,.65)),normalize(vec2(.83,-.26)),normalize(vec2(.13,-.83)),normalize(vec2(-.2,.55)));vec3 t(vec3 v){float f=0.,i=0.;for(int y=0;y<4;y++){float s=v.x*p[y].x+v.z*p[y].y;s=s*(4.-float(y)*.51013)+sp[162].z;float m=1./(float(y)+1.);f+=m*sin(s-.3*cos(s));i+=m*cos(s-.3*sin(s));}return vec3(f,0.,i);}bool t(vec3 f,vec3 e,vec3 v,float y,out float i){vec3 s=v-e;float m=dot(s,f);if(m<0.)return false;else{float n=length(s),o=n*n-m*m;if(o>y)return false;else{float t=sqrt(y-o);i=m-t;return true;}}}float t(vec2 f,out float v){vec4 s=texture(terrain,f/512.);v=s.y;return s.x*60.-12.1;}void t(vec3 f,vec3 y,out float v,out float i){vec3 s=(vec3(0.)-f)/y,o=(vec3(512.)-f)/y;v=max(min(s.x,o.x),min(s.z,o.z));i=min(max(s.x,o.x),max(s.z,o.z));}bool t(vec3 v,vec3 s,out float f,out vec3 i,out vec3 y,out float o,out float m){float x,e;t(v,s,x,e);if(e<x)return false;float z=max(0.,x);v=v+z*s;vec2 n,r,d;d=sign(s.xz);r=1./s.xz*d;n=r*(max(d,0)-fract(v.xz)*d);if(any(isinf(n)||isinf(r)))return false;vec2 c=floor(v.xz);float p=v.y,u=t(c,m);o=0.;for(f=0.;f<e-z;){vec2 k=vec2(float(n.x<n.y),float(n.x>=n.y));f=dot(n,k);float g=v.y+s.y*f;c=c+d*k;n=n+r*k;if(u>g)return i=vec3(.2,.071,.01)+step(27,u)*vec3(.1,-.06,0),o=1.2,y=vec3(0,1.,0),f=(u-v.y)/s.y,true;u=t(c,m);if(u>g){o=1.5;i=vec3(.2,.2,.2)+step(40,u)*vec3(0.,-.03,-.1);if(m==1.){float a=25.31;if(g<a)m=0.;}y=vec3(-d.x*k.x,0,-d.y*k.y);return true;}}return false;}float t(float f,vec3 v,vec3 s){float i=(1.-f)/(1.+f);i*=i;float y=-dot(v,s),o=1.-y,e=i+(1.-i)*o*o*o*o*o;return e;}const vec3 i=vec3(5e-06,1.5e-05,.00027)*15.,s=vec3(.00015,.00015,.00027)*15.;vec3 n(float f){return exp(-f*(i+s));}vec3 n(float f,float y){float v=.0003/16.*3.14159*(1.+y*y);vec3 o=1./(i+s)*(1.-exp(-f*s));float m=.476;vec3 x=vec3(.002,.0008,.0002)*(1.-m)*(1.-m)/(12.5664*pow(1.+m*m-2.*m*y,1.5));float e=20./(i.x+s.x)*(1.-exp(-f*s.x));return v*o+x*e;}void main(){vec3 s=normalize(vec3(1.,1.1,1.));vec2 o=2.*(gl_FragCoord.xy/y)-vec2(v/y,1.);vec4 i=cos(sp[161]),m=sin(sp[161]);mat3 x=mat3(i.y,0,-m.y,-m.x*m.y,i.x,-m.x*i.y,i.x*m.y,m.x,i.y*i.x);vec3 c=x*vec3(o,-2.),d=sp[160].xyz;c+=d;vec3 z=normalize(c-d);float r=1.;vec3 u=vec3(0);for(int p=2;p>0;p--){vec3 k,g,a,l;float w,b=0.,C=e;for(int h=0;h<f;h++){float F;if(t(z,d,sp[h*2].xyz,sp[h*2].w,F)){if(F<C)C=F,a=d+C*z,g=normalize(a-sp[h*2].xyz),l=sp[h*2+1].xyz,w=sp[h*2+1].w,b=t(w,g,z);}}float h;vec3 F,D;float q;if(t(d,z,h,F,D,w,q)){if(h<C){if(q==1.){float Z=25.91;vec3 Y=(d-vec3(386.,Z,447.))*512.;float X;if(t(Y,z*512,X,F,D,w,q))C=h+X/512.,l=F,l.z=l.z*2.;else{C=h*2.02;d=d+z*C;continue;}}else C=h,l=F;g=D;a=d+z*C*.9999;b=t(w,g,z);}}h=(-.5-d.y)/z.y;if(z.y<0.&&h<=C){a=d+z*h*.9999;g=vec3(0.,1.f,0.f)+t(a)*.01;g=normalize(g);b=t(1.1,g,z);vec3 Z=refract(z,g,1.-b);l=vec3(.05,.05,.15);if(t(a,Z*100,h,F,D,w,q))l+=F*exp(-h*40.);}k=reflect(normalize(z),g);if(C>=e){u+=n(C,dot(s,z))*r;break;}bool Z=t(a,s,h,F,D,w,q);if(!Z){for(int Y=0;Y<f;Y++){if(t(s,a,sp[Y*2].xyz,sp[Y*2].w,h)){Z=true;break;}}}if(!Z){float Y=dot(s,g);vec3 X=normalize(s-z);float W=pow(dot(g,X),121.);W=clamp(W,0.,1.);F=vec3(W)+l*Y;}else F=l*.02;F*=n(C);F+=n(C,dot(s,z));u+=F*r*(1.-b);r=r*b;z=k;d=a;}float g=min(sp[162].x,sp[162].y);if(g<=18)g=1.-g/8.;else g=0.;g=0.;float C=length(vec2(o.x*(y/v),o.y)),a=min(1.,smoothstep(.95*(1.-g/26.),1.31,C)*.6+g);vec3 Z=mix(u,vec3(0),a);fragColor=vec4(pow(u,vec3(1./2.2)),1.);}\0\0";
