+++ src/util.rs
@@ -4,6 +4,8 @@ use winapi::um::winnt::{
     FILE_APPEND_DATA,
     GENERIC_READ,
 };
+use core::panic::PanicInfo;
+
 
 #[cfg(feature = "logger")]
 use winapi::um::{
@@ -16,6 +18,31 @@ use winapi::um::{
     },
     handleapi::CloseHandle
 };
+
+#[panic_handler]
+#[no_mangle]
+pub extern fn panic( _info: &PanicInfo ) -> ! { loop {} }
+
+#[no_mangle]
+pub unsafe extern fn memset(dest: *mut u8, c: i32, n: usize) -> *mut u8 {
+    let mut i = 0;
+    while i < n {
+        *((dest as usize + i) as *mut u8) = c as u8;
+        i += 1;
+    }
+    dest
+}
+
+#[no_mangle]
+pub unsafe extern fn memcpy(dest: *mut u8, src: *const u8, n: usize) -> *mut u8 {
+    let mut i = 0;
+    while i < n {
+        *((dest as usize + i) as *mut u8) = *((src as usize + i) as *const u8);
+        i += 1;
+    }
+    dest
+}
+
 
 #[cfg(feature = "logger")]
 #[macro_export]
